+++ source/Properties/AssemblyInfo.cs
@@ -1,0 +1,26 @@
+﻿using System.Resources;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+
+// General Information about an assembly is controlled through the following 
+// set of attributes. Change these attribute values to modify the information
+// associated with an assembly.
+[assembly: AssemblyTitle("Lunatic Aeronautics Rover Anti Gravity System")]
+[assembly: AssemblyDescription("")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Lunatic Aeronautics")]
+[assembly: AssemblyProduct("Rover Anti Gravity System (RAGS)")]
+[assembly: AssemblyCopyright("© 2015 ximrm and Tuareg, © 2021,2022 zer0Kerbal")]
+[assembly: AssemblyTrademark("™ 2015 ximrm and Tuareg, ™ 2021,2022 zer0Kerbal")]
+[assembly: AssemblyCulture("")]
+
+// Setting ComVisible to false makes the types in this assembly not visible 
+// to COM components.  If you need to access a type in this assembly from 
+// COM, set the ComVisible attribute to true on that type.
+[assembly: ComVisible(false)]
+
+// The following GUID is for the ID of the typelib if this project is exposed to COM
+[assembly: Guid("3b23a104-0dee-435d-9441-b6397074edfc")]
+
+[assembly: NeutralResourcesLanguage("")]