+++ docs/changelog.md
@@ -1,4 +1,17 @@
-# Changelog  
+---
+permalink: /Changelog.html
+title: The Change Log
+description: The Opening Credits, and the closing credits, plus the first of two (or is three) end credit scenes
+tags: changes,changelog,change-log,page,kerbal,ksp,zer0Kerbal,zedK
+---
+
+<!-- 
+hdr-changelog.md v1.0.0.0
+Rover Anti Gravity System (RAGS)
+created: 13 May 2022
+updated:
+CC BY-ND 4.0 by zer0Kerbal
+--># Changelog  
   
 | modName    | Rover Anti Gravity System (RAGS)                                    |
 | ---------- | ------------------------------------------------------------------- |
@@ -190,7 +203,6 @@
 
 * 2015
 * Released for Kerbal Space Program 1.0.5
-* by ximrm and Tuareg
 
 ### Status